requirements.txt
google-genai
python-dotenv
imapclient
flask
